+++ lib/logger/logger.js
@@ -1,11 +1,10 @@
 'use strict'
 
 var request = require('request');
-var fs = require('fs');
-var querystring = require('querystring');
 var crypto = require('crypto');
 var path = require('path');
 var webwxbatchgetcontact = require('../webwx.js').webwxbatchgetcontact;
+var webwxgetmsgimg = require('../webwx.js').webwxgetmsgimg;
 
 var MSGTYPE_TEXT = require('../global.js').MSGTYPE_TEXT;
 var MSGTYPE_IMAGE = require('../global.js').MSGTYPE_IMAGE;
@@ -28,15 +27,6 @@ function wechatLogger(obj) {
 }
 
 
-function webwxgetmsgimg(msgId, obj, imgPath){
-  var imgUrl = `https://wx.qq.com/cgi-bin/mmwebwx-bin/webwxgetmsgimg?&MsgID=${msgId}&skey=${querystring.escape(obj.BaseRequest.Skey)}`;
-  // 保存图片到文件
-  try {
-    request.get(imgUrl, {jar: true}).pipe(fs.createWriteStream(imgPath));
-  } catch (e){
-    console.error('下载图像资源失败:', e);
-  }
-}
 
 function logImageMessage(o, obj) {
   var imgPath = path.join(process.cwd(), 'data/pic', crypto.createHash('md5').update(crypto.randomBytes(10)).digest('hex') + '.jpg');
