+++ lib/webwx.js
@@ -2,6 +2,10 @@
 var debug = (text)=>console.error("[DEBUG]", text);
 var inspect = require('util').inspect;
 var request = require('request');
+
+var querystring = require('querystring');
+var fs = require('fs');
+
 var cacheContact = require('./cache.js').cacheContact;
 
 var SPECIAL_USERS = require('./global.js').SPECIAL_USERS;
@@ -443,6 +447,16 @@ function webwxbatchgetcontact(username, obj) {
   });
 }
 
+function webwxgetmsgimg(msgId, obj, imgPath){
+  var imgUrl = `https://wx.qq.com/cgi-bin/mmwebwx-bin/webwxgetmsgimg?&MsgID=${msgId}&skey=${querystring.escape(obj.BaseRequest.Skey)}`;
+  // 保存图片到文件
+  try {
+    request.get(imgUrl, {jar: true}).pipe(fs.createWriteStream(imgPath));
+  } catch (e){
+    console.error('下载图像资源失败:', e);
+  }
+}
+
 function sessionStop(code, signal) {
   console.log('结束会话');
   process.exit(code);
@@ -461,3 +475,4 @@ module.exports.synccheck = synccheck;
 module.exports.webwxsync = webwxsync;
 module.exports.webwxsendmsg = webwxsendmsg;
 module.exports.webwxbatchgetcontact = webwxbatchgetcontact;
+module.exports.webwxgetmsgimg = webwxgetmsgimg;
